requirements.txt
fastapi
uvicorn
python-multipart
pydantic
requests
python-dotenv

